+++ packages/video_player/video_player_platform_interface/lib/method_channel_video_player.dart
@@ -113,7 +113,7 @@ class MethodChannelVideoPlayer extends VideoPlayerPlatform {
           return VideoEvent(
             eventType: VideoEventType.initialized,
             rotationCorrection:
-                (map['rotationCorrection'] as double?)?.toDouble() ?? 0.0,
+                (map['rotationCorrection'] as int?)?.toDouble() ?? 0.0,
             duration: Duration(milliseconds: map['duration']! as int),
             size: Size((map['width'] as num?)?.toDouble() ?? 0.0,
                 (map['height'] as num?)?.toDouble() ?? 0.0),
